# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: STACK Service Desk AI Solution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/reports/resolution-rate:
    get:
      operationId: getResolutionRate
      tags: [reports]
      summary: Resolution rate by use case
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resolution rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolutionRateReport"
  /v1/reports/sla-compliance:
    get:
      operationId: getSlaCompliance
      tags: [reports]
      summary: SLA compliance report
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: SLA compliance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaComplianceReport"
  /v1/reports/ticket-trends:
    get:
      operationId: getTicketTrends
      tags: [reports]
      summary: Ticket volume trends
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: granularity
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Ticket trends data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketTrendsReport"
  /v1/reports/ai-accuracy:
    get:
      operationId: getAiAccuracy
      tags: [reports]
      summary: AI confidence and accuracy analytics
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: AI accuracy data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAccuracyReport"
  /v1/reports/agent-performance:
    get:
      operationId: getAgentPerformance
      tags: [reports]
      summary: Agent performance metrics
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Agent performance data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentPerformance"
...
  /v1/sops:
    get:
      operationId: listSops
      tags: [sops]
      summary: List all SOPs
      responses:
        "200":
          description: SOPs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sop"
    post:
      operationId: createSop
      tags: [sops]
      summary: Upload new SOP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SopInput"
      responses:
        "201":
          description: SOP created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sop"
  /v1/sops/{id}:
    get:
      operationId: getSop
      tags: [sops]
      summary: Get SOP detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SOP detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sop"
    patch:
      operationId: updateSop
      tags: [sops]
      summary: Update SOP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SopUpdate"
      responses:
        "200":
          description: SOP updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sop"
    delete:
      operationId: deleteSop
      tags: [sops]
      summary: Deactivate SOP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SOP deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/sops/search:
    post:
      operationId: searchSops
      tags: [sops]
      summary: Semantic search SOPs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SopSearchInput"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sop"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    ResolutionRateItem:
      type: object
      required: [use_case, auto_count, manual_count, total, auto_rate]
      properties:
        use_case:
          type: string
        auto_count:
          type: integer
        manual_count:
          type: integer
        total:
          type: integer
        auto_rate:
          type: number
    ResolutionRateReport:
      type: object
      required: [items, date_from, date_to]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResolutionRateItem"
        date_from:
          type: string
        date_to:
          type: string
    SlaComplianceReport:
      type: object
      required: [met_count, breached_count, at_risk_count, compliance_percent, trend]
      properties:
        met_count:
          type: integer
        breached_count:
          type: integer
        at_risk_count:
          type: integer
        compliance_percent:
          type: number
        trend:
          type: array
          items:
            type: object
    TicketTrendsReport:
      type: object
      required: [trend]
      properties:
        trend:
          type: array
          items:
            type: object
    AiAccuracyReport:
      type: object
      required: [avg_confidence, distribution, by_use_case]
      properties:
        avg_confidence:
          type: number
        distribution:
          type: array
          items:
            type: object
        by_use_case:
          type: array
          items:
            type: object
        trend:
          type: array
          items:
            type: object
    AgentPerformance:
      type: object
      required: [agent_id, full_name, tickets_handled, avg_resolution_mins]
      properties:
        agent_id:
          type: string
        full_name:
          type: string
        tickets_handled:
          type: integer
        avg_resolution_mins:
          type: number
...
    Sop:
      type: object
      required: [sop_id, title, use_case, content, version, is_active, created_at]
      properties:
        sop_id:
          type: string
        title:
          type: string
        use_case:
          type: string
          enum: [sharepoint_access, sharepoint_admin, license_bluebeam, license_adobe, license_o365, dl_update, windows_troubleshooting]
        content:
          type: string
        version:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    SopInput:
      type: object
      required: [title, use_case, content]
      properties:
        title:
          type: string
        use_case:
          type: string
          enum: [sharepoint_access, sharepoint_admin, license_bluebeam, license_adobe, license_o365, dl_update, windows_troubleshooting]
        content:
          type: string
        version:
          type: string
    SopUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        version:
          type: string
        is_active:
          type: boolean
    SopSearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        use_case:
          type: string
...