# lib/api-zod/src/generated/types/agentPerformance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 1.0.0
 */

export interface AgentPerformance {
  agent_id: string;
  full_name: string;
  tickets_handled: number;
  avg_resolution_mins: number;
}