# lib/api-zod/src/generated/types/sopSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 1.0.0
 */

export interface SopSearchInput {
  query: string;
  use_case?: string;
}