# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 1.0.0
 */

export interface SuccessResponse {
  success: boolean;
  message: string;
}